In counter.ts, I created ngOnInit event. This event is useful for determining what should be rendered after the component has been initialized.
It runs once after rendering and is useful for asynchronous tasks, such as handling promises(then) and petitions (subs). This is particularly helpful for tasks that take time, such as connecting to an API or a database, and then rendering the component.
It cant run twice.

To identify the event, I used console.log;
I printed the state of the inputs (duration, message) using => this.duration