In wave-audio.ts I imported the WaveSurfer library and used the ngAfterViewInit event.
To use the library I needed the ngAfterViewInit event.

To determine if the child component have already been rendered I used ngAfterViewInit event. This allow me to set up WaveSurfer to provide the audio url and container.

This url comes from a server or the assets of our project.

Regarding the url, I wanted this component to be reusable, so this url will be sent as an input. It will be mandatory and named audioUrl.
Its good practice to initialize  a variable.
AudioUrl: string = ‘’;
However, sometimes its useful to initialize a variable later when needed. To do this, in typescript you cand add the ! sign .