Regarding  the container, I had to set up the space where the audiogram will be generated. First, I worked on wave-audio.html and then on wave-audio.ts.
In wave-audio.html I used an element reference -# ( wich is very common in javascript)- to know when this element is rendered, add properties to this element and to use it to set up the container for WaveSurfer

To get this reference in wave-audio.ts I used the @ViewChild decorator, who calls to wave reference and named container. This is an Html element so I wrote ElementRef, which comes from @angular/core. This element is not initialized immediately so I added the ! sign

To set up the container for WaveSurfer I made sure to use its native element

To set up a container in Javascript, You would use document.getElementbyId(‘container’) to render in the container.