Adding the play icon
In wave-audio.ts  I created the playPause event, so I needed a reference to the audio. To do this, I created an instance of WaveSurfer named ws, set it up as private and it will be initialized later.
After that, in the event, I executed the PlayPause method of this library.

In wave audio.html I executed the playPause event. I used a button named play and the click attribute.