Adding the play/pause icon
in wave-audio.ts, first, I determined the state of the audio from WaveSurfer. I used a signal named isPlaying and set it to false because initially the audio does not play.

To determine if the audio is playing, I used the on method  of this library, setting it up for the play and pause events:
1) When the play event is called, I used an arrow function to change the isPlaying state to true.
2) When the pause event is called,  I used an arrow function to change the isPlaying state to false.

In wave-audio.html I used the initial state of the audio to change the botton text throught *ngIf and else. Dont forget the initial state is paused.
I used a span for Pause and ng-template for Play:

1) In span for Pause: I used *ngIf to determine if audio is playing. Since the initial state of the audio is not playing, it will show Pause in the layer, otherwise it will use the elseBlock.
2) Regarding Play, I used the ng-template to execute the #elseblock reference to display Play in the layer.