I realized that the list component is a page that renders some products. The list.html initially had an app-product component and a list of products hardcoded manually, so It was time to automate this process.
In list.ts, I created a signal that contains a list of Product. To achieve this, I created a variable named products. This signal is an array of Product initialized throught the contructor.

I needed a interface to define the elements for a Product. To do this, I created the product.model.ts file in the path shared/models. There I defined and exported an interface named Product, with attributes of a Product such as: id, title, price, image. After this, to use this interfaz so I imported this interfaz into the list component to use it

In list.ts I imported the Product interface from the models folder.

Regarding the signal, it was initialized with initial products throught the constructor. Here I harcoded the initial products.

In the constructor I used the const initProducts , to define initial products, This is an array of Product[], where I specified the values of all attributes of a Product to create the first Product. I then defined the values of 2 products.

in the constructor, the initProducts array was assignated to the products signal.

In this way, I initialized a set of products in the constructor.

In list.html, I used the products signal with the ngfor directive to render the products dinamically. I used the iterator variable named prod for the signal named products. Here, this signal was executed.