Optimizing the code for the last client request The changes allow the new attribute creationAt to appear as a part of the product details.
In list.html, Instead of sending each attribute individually, I sent the entire product object. To achieve this, I used the prod iterator as a new attribute, wich is generic and receives the entire Product object

In product.ts I set up this new input named prod , to receive the entire Product object. To do this, I used the Product interface to type the new input named prod.
Since this input was not initialized immediately, I added the ! Sign  to indicate that it will be initialized later.

In product.html I updated the template to use the new prod input, accessing the attributes of the Product object  directly throught this input.