I created the header for the application
To do this, I used the flowbite website to find a template for the navbar and an online icon to use it as a header.

I created a new component named header by writing this command in the console: ng g c domains/shared/components/header

In header.html I used the template from Flowbite website.

In the list.html. to display the header in the application, I used the header selector app-header.
In list.ts, to make the header selector functional, I imported the header component.