Shopping cart implementation
I planned that when clicking on cart icon, a side menu with a list of favorite products would be displayed

In header.html I added a template from FlowByte website for this side menu.

To display or hide  this side menu I set the translate-x-full class dinamically from its template. To do this, In header.ts I created a signal named hideSideMenu, defaulting to true. In this way, the side is initially hidden.
Also, I created a method named toogleSideMenu to change the state of signal to hide or display the shopping cart side menu . To do this, I updated the signal.

In the header.html I executed the method named toogleSideMenu. To do this, I changed the li tag to button tag for the shopping cart icon. Then, I added a click event to execute this method in both buttons: 1) shopping cart icon; 2) side menu

Finally, I set the translate-x-full class dynamically. To do this, I used the dinamyc class by property biding and  subscribed it to the signal named hideSideMenu.