ADDING PRODUCTS TO THE SHOPPING CART SIDE MENU
The objective is to allow users to add products to the shopping car side menu when clicking on the Add to Cart button. To achieve this I used reactivity and communication between parent and child components.

CREATING a signal in list.ts.
I created a signal named cart to save the products that the user has selected. This signal is an array of Product.

COMPONENT STRUCTURE
Before to continue, its time to remember that:
The parent component is list.ts
The child component is product.ts

MODIFYING THE EVENT IN list.ts.

There was an event named fromChild, wich was a method to receive information from child component in parent component. I modified this event to addToCartMenu to notify that a product has been added to the shopping car side menu.

UPDATING list.html
I executed the addToCartMenu event because the fromChild event is not available.

EMITTING FROM product.ts
There was an event addCartHandler to emit a message from the child component to parent component. This message was initially a string. I changed this message to emit the entire product

HANDLING THE EVENT in list.ts
The addToCartMenu method previously received a string (a message) from the addCartHandler event. I modified this method to receive a Product object, indicating the product being added to the shopping cart

I added the cart signal to addCartMenu method. I updated this signal as a list, transitioning from previous state (a empty list) to new state (new list containing both any previously products and the newly selected products). Finally, I changed the event parameter to product in this method.

DISPLAYING THE CART COUNT IN list.html
I added an h3 tag of HTML to show how many products are in the shopping cart. To do this, I subscribed to the cart signal and displayed its length.
I modified the style of the h3 tag to make it more visible.