SENDING PRODUCTS TO THE SHOPPING CART SIDE MENU
1. Reception of products in child component (header.ts)

I created an input named cartSide in the header.ts because the comunication is from the parent to the child component.
This input is set to be required and will receive a list of products from the parent component. So the input is an array of Product and its initialized as an empty array.
The list of products sent from the parent component consists of those selected by the user, which have been added to the shopping cart side menu.
Finally, I imported CommonModule in the header.ts to use the cardSide input in the layout.

2. Sending information from the parent component (list.html)

I subscribed the cartSide input to the cart signal.
The cartSide input was created in the child componet, header.ts
The cart signal was created in fthe ather component, list.ts.

3. Displaying the products in the shopping cart side menu (header.html)

I used a div tag to display the products. I used *ngfor to render the products from the cartSide input
Inside the div,I used a p (paragraph) tag to display the caracteristics of these products. I used string interpolation {{}} to print these caracteristics.