STYLING THE CART COUNT in header.html
1. Adding the cart count to the cart button

I added the cart count inside the cart button. To do this: I used a div tag and string interpolation for the input cartSide to displayed its length.
Finally, I deleted the subscription to the cart signal in list.html because I have the cart count inside the cart button.

2. Changing the style of the button

I modified the padding, border and added a relative style. Specifically I added p-2 for 8 pixels of padding; border rounded: and relative.

3. Modifying the Style of the Cart Count:

I styled the cart count. I added: an absolute style; rounded bord; -top- 2 and -left-2 (to position); h-5 w-5 (to set the height and width); bg-black (to change the background color to black); text-white (to modified the font color to white); text-xs (to modified font size to extra small).