Improving the performance
In header.ts:

Executing the calculateTotal method directly in the header layout is not a good practice because it decreases performance. Only a signal should be executed in the layout, not a function or method.
If the cardSide were a signal the solution would be to use a computed property, which is calculated from other signals. However since the cartSide is an input since, a different approach is needed.

The best solution is to calculate the total each time the cartSignal input receives a new value. To do this:
1. I created a new signal named total and initialized it to zero.
2. I created a new method named ngOnChanges to set the total signal each time the input receives a new value by using  the calculateTotal method.
3. Validating. Inside the ngOnChanges I validated if cartSide has actually changed before recalculating the total.
3. In the header layout, the total signal is excuted instead of the calculateTotal method.