Avoiding the prop drilling  or input drilling in Angular.
Prop drilling is a complex problem when the communication between components involves three or more levels resulting in many nodes to transfer information from a place of one extreme of the tree to the other.

1 Creating a service to facilitate the communication between components.

To solve the problem of prop drilling, it is common to use a store to make the project easier to maintain

Component interaction
A component saves data in the store, or state Managment (a service in angular)to communicate with other components.

Subscriptions
If a component is interested in a specific state then it will subscribe to a signal or observable. This signal exists globally in the store

State update
Once a component subscribes to the store, it receives a notifications about any state changes, allowing it to decide what to do (e.g rendering, updating a view, etc.

1.1 Creating and using the service

A service is commonly shared then to create it, I used the following command in the console to create the cartsideService:
ng g s domains/shared/services/cartsideService

Service responsabilities
A service handles business logic ; data manipulation; state managment; communication with external service (e.g. REST API)

It is not used for rendering wich is the responsibility of the component