PROP DRILLING- PART 2
2. Handling the state of the shopping cart in cartsideservice.ts

This service handles the state of the shopping cart through a signal. It comprises the Injectable decorator to inject dependencies and a  class named CartsideService.

2.1 Creating the signal for components.

In the cartside.service.ts:

Inside the CartsideService class, I created a signal named cart to be used by components. Components can subscribe to this signal to facilitate the communication between them.
This signal handles the products in the shopping cart.

The cart signal is an array of Product and is initialized as an a empty array.