PROP DRILLING- PART 3 3. Calculating the total price of the products added to the shopping cart.
In cartsideservice.ts

Initially the total price of the products added to shopping cart was calculated in the header.ts.
However, to improve the structure and maintainability of the code, I moved this calculation to cartsideService.ts.

3.1 Steps to calculate the Total:
3.1.1 Creation of the total computed variable:
In cartside.Service.ts
There is a cart signal is used to manage the list of products in the shopping cart. I can leverage this signal to calculate the total price through creation the computed variable named total.
The computed variable is a signal calculated from others signals.

3.1.2 Setting up the total computed variable

I set up the total computed variable as an array. First I added the list of products that are in the shopping cart through an subscription to cart signal. Second, this total computed variable will return the total price of all products added through a JavaScript arrow function named reduce of the cart signal.

The reduce function has an accumulator variable named total and the current productSide being processed. This function adds the current total plus productSide.price. Finally the accumulator variable  is initialized to zero.

Finally the state management for the shopping cart is in the store and it is responsible for  communicating the state of the shopping cart to the interested components