PROP DRILLING – PART 4 4. Subscription to the signal of the service
In header.ts, I deleted an input named cartSide, which it communicated from the child component to the parent component (list.ts). I deleted it because the cartsideService now handles this communication.

4.1 Injection of dependencies to allow communication from the child componet to the parent component

In header.ts, to allow the cartsideService to handle the comnumication between components, I injected this service by creating a private variable named cartsideService. It is private variable to avoid rendering it in the component.

To set the cartsideService variable I used the inject method to call to the CartsideService.
This service was previously imported into the header.ts component.