PROP DRILLING – PART 4 – 4.2 4.2 New setting of cart and total signals of the service
The cartsideService comprises a addtoCart method and two signals: total and cart (it is a list of products added to the shopping cart)

In the header.ts there was a signal named total. This total was calculated by ngOnChanges and used to compute the total price of products. This signal can now be directly utilized and delegateS to the service. To do this, I set this total variable to call the cartsideService, and the total signal.
Similarly I set the cart variable to call the cartsideService and the cart signal.

These 2 variables are signals and will be available to receive subscriptions

Finally It is possible delete the ngOnchanges and calculateTotal methods.

In this way the service handles the data management and the components handle rendering.