PROP DRILLING – PART 4 – 4.4 4.4 Injection of dependencies to allow the communication from the parent component to the child component
In list.html component:
In the app-header section, this layout was sending information from the parent component (list.html) through an subscription of the cartSide input to the cart signal. This subscription was deleted because there is a service to manage the communication between components.

In list.ts:
There was the cart signal to send information from the parent components. It was deleted because the CartsideService is now available . To do this, I did the injection of dependencies using a private cartsideService variable, in the same way I did in header.ts.