PROP DRILLING – PART 4 – 4.5
4.5 Adding products to the shopping cart
In list.ts:
To add a product to the shopping cart I modified the addToCart method to use the cartsideService and call its method named addToCart for adding product.

Finally, the general state of the shopping cart is managed by cartsideService. The communication between components is handled by a service and not by inputs or outputs.