ProductService CONFIGURATION TO PERFORM CRUD OPERATIONS.
1.4 Accesing the API

In app.config.ts

 - Importing mondule
I imported the provideHttpClient new module from @angular/common/http to allow the ProductService (product.service.ts) access this module

 - Provider configuration

The provideHttpCliente method is called within the array of providers to allow the ProductService makes API calls

1.5 Connection to API

In product.service.ts

- Dependency injection

Once the service boilerplate was created. I performed dependency injection to use HttpClient. To do this, I created a private variable named http within the ProductService class and injected HttpClient.

- HttpClient usage

HttpClient allows the connection to the API to send, request, get and process information.

- Getting products

Inside the ProductService class, I created a method named getProducts to retrieve products. To do this, this method is typed to return an array of Product and uses the http variable to connect to the API.
The http.get is called to connect to the products endpoint (url) and retrieve the products.