1. PIPE
A pipe transforms data by taking input and converting it into a different format using its | symbol.

1.1 Using pipe

Pipes are primarly used in templates rather than in business logic.

In productt.html:

The | uppercase pipe is used to transform the title to uppercase.

The | currency pipe is used to format money, and the | date is used to format dates.

For more types of pipes  refer to the Angular documentation.

1.2 Import CommonModule
In product.ts:

The CommonModule is imported to ensure proper functionality of pipes.