PIPE CREATION
1. New pipes

Reverse pipe was created using the comand in the console

ng g p domains/shared/pipes/reverse

Time-ago pipe was the second pipe created using the same comand in the console

2. Reverse pipe

In reverse.pipe.ts:

There is a @Pipe decorator named reverse.

Reverse pipe was created to reverse the letters in a title product.

The pipe acts as transformer, taking an input string, transforming it, and returning a reversed string.

The reverse pipe receives a string, without arguments, and return a string. To set the return, It uses split (to convert a string into an array), reverse (to reverse the elements in an array), and join (to add all elements into a string).

3. Using the reverse pipe

In product.ts:
ReversePipe is imported

In product.html:
Reverse pipe is applied alongside the uppercase pipe