1. HANDLING DATE 1.2 Setting timeAgo pipe
In time-ago.pipe.ts:

Time-ago.pipe,ts file was created to display the time elapsed since creation date.

Here, there is a @Pipe decorator named timeAgo, which is used in the layout (to use in the template)

To set this pipe, I imported the formatDistance function from the newly installed library.

This pipe acts as transformer, taking an input string, transforming it, and returning the result of the comparison between today’s date and the creation date. To achieve this, I followed these steps:

- Set the time-ago pipe to receive a string (without arguments) and return a string.
- Create the creation date (creationDate) convertin the input string (value) to date  using new Date()
- Create the today’s date (currentDate) using new date().
- Use the formatDistance function to return the time difference between currentDate and creationDate.

1.3 Using the timeAgo pipe

In product.ts:
Import TimeAgoPipe

In product.html:
Apply the timeAgo pipe alongside prod.creationAt