2. DIRECTIVES
The directives are used to directly manipulate the DOM

I created a directive to change the background to red in a short space of the about layout.

2.1 Creating the directive

The highlight directive is created using the following command in the console:

ng g d domains/shared/directives/highlight

2.2 Setting the directive

In highlight.directive.ts:

Here, there is a decorator named directive, which includes a selector.
The highlight selector was renamed for easy use in the template (layout)

Its necesesary an element or attribute and its injection of dependence to apply a directive. . To do this, I created element to inject a dependence (ElementRef).

The DOM is modified using ngOnInit lifecycle event. Once initialized, calls to element to access to native element and finally to change its style and background to yellow using Javascript.

2.3 Using the directive

In about.ts:
The HighlightDirective class is imported from the newly created directive

In about.html:
The span tag is used to apply the directive using its selector named highlight