CREATING A 404 ROUTE
The objectiv is to display an error when the user accesses a route that is not available.

1. I created the not-found page using this command in the console to generate the component:
ng g c domains/info/pages/not-found

2. Create a layout for the 404 route

In nout-found.html, I designed the layout with a link to navigate back home.

3. Set up the route

In app.routes.ts:
I added a new path to connect the Not Found component.
I used ** as a wildcard to catch undefined routes.