routerLink - Single page application optimization
Objective:
Optimize the 404 error page by preventing a full page reload when navigating back to the home page.

Original issue
The BackToHome button on the 404 error page was using an Href tag, which caused the entire page  to reload when navigating back to the home page.

Solution:
To achieve a smoother user experience, I replaced the Href tag with Angular’s routerLink directive, enabling single-page application (SPA) navigation.

In not-found.html:
Replaced the href tag with routerLink to redirect the user to the home page without triggering a full page reload.

In not-found.ts:
Imported Angular’s RouterLinkWithHref directive from the router package. This allows the BackTo Home button to use Angular’s internal routing mechanism.

Result:
The BackToHome button now redirects the user to the home page without reloading the entire page, improving the efficiency and user experience of the application.